transport complited

+++ src/app/components/courier/all-couriers/all-couriers.component.ts
@@ -6,6 +6,7 @@ import { ButtonModule } from 'primeng/button';
 import { TableModule } from 'primeng/table';
 import { RouterLink } from '@angular/router';
 import { TitleService } from '../../../services/title.service';
+import { ConfirmationService, MessageService } from 'primeng/api';
 
 @Component({
   selector: 'app-all-couriers',
@@ -23,7 +24,10 @@ export class AllCouriersComponent implements OnInit {
 
   constructor(
     private readonly courierService: CourierService,
-    private readonly titleService: TitleService
+    private readonly titleService: TitleService,
+    private readonly messageService: MessageService,
+    private readonly confirmationService: ConfirmationService,
+
   ) {}
 
   ngOnInit(): void {
@@ -34,6 +38,19 @@ export class AllCouriersComponent implements OnInit {
 
   delete(courier : ICourier){
     console.log(courier);
-    this.courierService.remove(courier.id);
+
+    this.confirmationService.confirm({
+      header: 'Confirmation',
+      message: 'Are you sure you want to proceed with the delete?',
+      accept: () => {
+        this.courierService.remove(courier.id);
+        this.messageService.add({ 
+            severity: 'success',
+            summary: 'Your courier has been deleted',
+            detail: '',
+            icon: 'pi pi-check'
+         })
+      },
+    })
   }
 }
